requirements.txt
openai==0.26.1
pinecone_client==2.1.0
streamlit==1.18.0
docx2txt
pypdf==3.4.1
langchain==0.0.91
boto3==1.20.22

